requirements.txt
aiogram >= 3.1.1
fastapi
phonenumbers
uvicorn
python-decouple
urllib3
requests
httpx
apscheduler
pydantic
python-digest
aiohttp-auth
websocket-client
pyyaml